fpsensor: Fix signed/unsigned comparison in FpFrameSizeCache
The variable `capture_type` (an enum) was being directly compared
against `frame_sizes_.size()` (a size_t, which is unsigned). This
triggered a compilation error when `-Werror=sign-compare` was enabled.

Explicitly cast `capture_type` to `size_t` for the bounds check to
resolve the signed/unsigned mismatch.

BUG=b:457818548
TEST=CQ
LOW_COVERAGE_REASON=b:457818548, test will be added later

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7122721
Tested-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>

+++ common/fpsensor/fpsensor_frame_size.cc
@@ -73,7 +73,8 @@ error_exit:
 uint32_t
 FpFrameSizeCache::get_frame_size(enum fp_capture_type capture_type) const
 {
-	if (capture_type >= frame_sizes_.size() || capture_type < 0) {
+	if (static_cast<size_t>(capture_type) >= frame_sizes_.size() ||
+	    capture_type < 0) {
 		CPRINTF("Error: Invalid fp_capture_type %d requested (max: %zu), "
 			"returning size 0.",
 			capture_type, frame_sizes_.size());
